requirements.txt
asgiref==3.8.1
black==21.10b0
blinker==1.4
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.4
djangorestframework==3.15.2
drf-nested-routers==0.94.1
filelock==3.6.0
gyp==0.1
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.6
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
oauthlib==3.2.0
pathspec==0.9.0
pillow==10.3.0
platformdirs==2.5.1
psutil==6.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
tomli==1.2.2
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
virtualenv==20.13.0+ds
wadllib==1.3.6
zipp==1.0.0
